refactor: api 메서드 파일로 분리
